+++ src/devices/centralite.ts
@@ -6,7 +6,8 @@ import * as constants from "../lib/constants";
 import * as exposes from "../lib/exposes";
 import * as m from "../lib/modernExtend";
 import * as reporting from "../lib/reporting";
-import type {DefinitionWithExtend, Fz} from "../lib/types";
+import type {DefinitionWithExtend, Fz, Tz} from "../lib/types";
+import * as utils from "../lib/utils";
 
 const e = exposes.presets;
 const ea = exposes.access;
@@ -32,6 +33,33 @@ export const centraliteExtend = {
             commands: {},
             commandsResponse: {},
         }),
+};
+
+const tzLocal = {
+    light_onoff_restorable_brightness: {
+        /**
+         * Some devices reset brightness to 100% when turned on, even if previous brightness was different
+         * This uses the stored state of the device to restore to the previous brightness level when turning on
+         */
+        key: ["state", "brightness", "brightness_percent"],
+        options: [exposes.options.transition()],
+        convertSet: async (entity, key, value, meta) => {
+            const deviceState = meta.state || {};
+            const message = meta.message;
+            const state = utils.isString(message.state) ? message.state.toLowerCase() : null;
+            const hasBrightness = message.brightness != null || message.brightness_percent != null;
+
+            // Add brightness if command is 'on' and we can restore previous value
+            if (state === "on" && !hasBrightness && utils.isNumber(deviceState.brightness) && deviceState.brightness > 0) {
+                message.brightness = deviceState.brightness;
+            }
+
+            return await tz.light_onoff_brightness.convertSet(entity, key, value, meta);
+        },
+        convertGet: async (entity, key, meta) => {
+            return await tz.light_onoff_brightness.convertGet(entity, key, meta);
+        },
+    } satisfies Tz.Converter,
 };
 
 export const fzLocal = {
@@ -109,7 +137,7 @@ export const definitions: DefinitionWithExtend[] = [
         vendor: "Centralite",
         description: "3-Series smart dimming outlet",
         fromZigbee: [fzLocal.restorable_brightness, fz.on_off, fz.electrical_measurement],
-        toZigbee: [tz.light_onoff_restorable_brightness],
+        toZigbee: [tzLocal.light_onoff_restorable_brightness],
         exposes: [e.light_brightness(), e.power(), e.voltage(), e.current()],
         configure: async (device, coordinatorEndpoint) => {
             const endpoint = device.getEndpoint(1);
@@ -149,7 +177,7 @@ export const definitions: DefinitionWithExtend[] = [
         vendor: "Centralite",
         description: "3-Series smart dimming outlet",
         fromZigbee: [fzLocal.restorable_brightness, fz.on_off, fz.electrical_measurement],
-        toZigbee: [tz.light_onoff_restorable_brightness, tz.ignore_transition],
+        toZigbee: [tzLocal.light_onoff_restorable_brightness, tz.ignore_transition],
         exposes: [e.light_brightness(), e.power(), e.voltage(), e.current()],
         configure: async (device, coordinatorEndpoint) => {
             const endpoint = device.getEndpoint(1);
